refactor: update dir structure and move command logic into function
