+++ src/commands/apextests/list.ts
@@ -1,25 +1,13 @@
-/* eslint-disable no-console */
 'use strict';
-/* eslint-disable no-await-in-loop */
 
 import { SfCommand, Flags } from '@salesforce/sf-plugins-core';
 import { Messages } from '@salesforce/core';
 
-import { getPackageDirectories } from '../../helpers/getPackageDirectories.js';
-import { validateTests } from '../../helpers/validateTests.js';
-import { extractTypeNamesFromManifestFile } from '../../helpers/parsers.js';
-import { formatList } from '../../helpers/formatters.js';
-import { searchDirectoryForTestClasses, searchDirectoryForTestNamesInTestSuites } from '../../helpers/readers.js';
-import { SearchResult } from '../../helpers/types.js';
+import { listTests } from '../../core/listTests.js';
+import { ApextestsListResult } from '../../utils/types.js';
 
 Messages.importMessagesDirectoryFromMetaUrl(import.meta.url);
 const messages = Messages.loadMessages('apextestlist', 'apextests.list');
-
-export type ApextestsListResult = {
-  tests: string[];
-  command: string;
-  // TODO: in the future, return the test suites as well
-};
 
 export default class ApextestsList extends SfCommand<ApextestsListResult> {
   public static readonly summary = messages.getMessage('summary');
@@ -32,6 +20,7 @@ export default class ApextestsList extends SfCommand<ApextestsListResult> {
       description: messages.getMessage('flags.format.description'),
       char: 'f',
       required: false,
+      default: 'sf',
     }),
     manifest: Flags.string({
       summary: messages.getMessage('flags.manifest.summary'),
@@ -62,82 +51,19 @@ export default class ApextestsList extends SfCommand<ApextestsListResult> {
     }),
   };
 
+
   public async run(): Promise<ApextestsListResult> {
     const { flags } = await this.parse(ApextestsList);
 
-    const format = flags.format ?? 'sf';
-    const manifest = flags.manifest ?? undefined;
-    const ignoreMissingTests = flags['ignore-missing-tests'] ?? false;
-    const ignoreDirs = flags['ignore-package-directory'] ?? [];
-    const noWarnings = flags['no-warnings'];
+    const result = await listTests({
+      format: flags.format,
+      manifest: flags.manifest,
+      ignoreMissingTests: flags['ignore-missing-tests'],
+      ignoreDirs: flags['ignore-package-directory'],
+      noWarnings: flags['no-warnings'],
+      warn: this.warn.bind(this),
+    });
 
-    let testClassesNames: string[] | null = null;
-    const testSuitesNames: string[] = [];
-    const allTestClasses: string[] = [];
-    const warnings: string[] = [];
-
-    // Get package directories full paths
-    const packageDirectories = await getPackageDirectories(ignoreDirs);
-
-    if (manifest) {
-      const manifestMetadata = await extractTypeNamesFromManifestFile(manifest);
-      testClassesNames = manifestMetadata.filter((name) => !name.endsWith('testSuite-meta.xml'));
-    }
-
-    // Loop through each directory and search for test classes
-    for (const directory of packageDirectories) {
-      const searchResult: SearchResult = await searchDirectoryForTestClasses(directory, testClassesNames);
-      allTestClasses.push(...searchResult.classes);
-      testSuitesNames.push(...searchResult.testSuites);
-      if (searchResult.warnings?.length) {
-        warnings.push(...searchResult.warnings);
-      }
-    }
-
-    if (testSuitesNames.length > 0) {
-      for (const directory of packageDirectories.filter((dir) => dir.includes('testSuites'))) {
-        const testNames: string[] = await searchDirectoryForTestNamesInTestSuites(directory, packageDirectories);
-        allTestClasses.push(...testNames);
-      }
-    }
-
-    let finalTestMethods = Array.from(new Set(allTestClasses.map((test) => test.trim())));
-
-    // If ignore-missing-tests is true, validate the test methods
-    if (ignoreMissingTests) {
-      const { validatedTests, warnings: validationWarnings } = await validateTests(
-        finalTestMethods,
-        packageDirectories,
-      );
-      finalTestMethods = validatedTests;
-
-      if (validationWarnings.length > 0) {
-        warnings.push(...validationWarnings);
-      }
-
-      if (validatedTests.length === 0) {
-        warnings.push('No test methods declared in your annotations were found in your package directories');
-      }
-    }
-
-    // Print all collected warnings
-    if (!noWarnings && warnings.length > 0) {
-      warnings.forEach((warning) => {
-        this.warn(warning);
-      });
-    }
-
-    finalTestMethods.sort((a, b) => a.localeCompare(b));
-
-    if (finalTestMethods.length === 0) {
-      this.warn('No test methods found');
-      return {
-        tests: [],
-        command: '',
-      };
-    }
-
-    const result = await formatList(format, finalTestMethods);
     this.log(result.command);
     return result;
   }
