--- src/helpers/constants.ts
@@ -1,3 +1,0 @@
-'use strict';
-export const SFDX_PROJECT_FILE_NAME = 'sfdx-project.json';
-export const SEARCHABLE_METADATA_FOLDERS = ['classes', 'triggers', 'testSuites'];
