+++ src/utils/getPackageDirectories.ts
