+++ src/utils/matchWildcard.ts
@@ -1,0 +1,4 @@
+export function matchWildcard(wildcard: string, test: string): boolean {
+  const regex = new RegExp(`^${wildcard.replace('*', '.*')}$`);
+  return regex.test(test);
+}
